+++ postFilters/default_f.go
@@ -1,7 +1,11 @@
 package postFilters
 
 import (
+	"fmt"
+
 	"github.com/Laisky/go-fluentd/libs"
+	"github.com/Laisky/go-utils"
+	"go.uber.org/zap"
 )
 
 type DefaultFilterCfg struct {
@@ -21,12 +25,23 @@ func NewDefaultFilter(cfg *DefaultFilterCfg) *DefaultFilter {
 	}
 }
 
+const DefaultSearchField = "message"
+
 func (f *DefaultFilter) Filter(msg *libs.FluentMsg) *libs.FluentMsg {
 	for k, v := range msg.Message {
 		switch v.(type) {
 		case []byte: // convert all bytes fields to string
 			msg.Message[k] = string(v.([]byte))
 		}
+	}
+
+	// DefaultSearchField must exists, let kibana can display this msg
+	switch msg.Message[DefaultSearchField].(type) {
+	case string:
+	case nil:
+		msg.Message[DefaultSearchField] = ""
+	default:
+		utils.Logger.Error("unknown message type", zap.String("message", fmt.Sprintf("%+v", msg.Message[DefaultSearchField])))
 	}
 
 	return msg